Refactor Spotify Metadata Cache
We will soon have an Apple Music Metadata cache and a lot of the code can
be shared between the two. Hence, refactoring.
