+++ docker/services/spotify_metadata_cache/consul-template-spotify-metadata-cache.conf
@@ -4,7 +4,7 @@ template {
 }
 
 exec {
-    command = "run-lb-command python3 -u -m listenbrainz.spotify_metadata_cache.spotify_metadata_cache"
+    command = "run-lb-command python3 -u -m listenbrainz.metadata_cache.spotify.runner"
     splay = "60s"
     reload_signal = "SIGHUP"
     kill_signal = "SIGTERM"
