+++ listenbrainz/metadata_cache/spotify/runner.py
@@ -1,0 +1,11 @@
+from listenbrainz.metadata_cache.consumer import ServiceMetadataCache
+from listenbrainz.metadata_cache.spotify.handler import SpotifyCrawlerHandler
+from listenbrainz.webserver import create_app
+
+
+if __name__ == "__main__":
+    app = create_app()
+    with app.app_context():
+        handler = SpotifyCrawlerHandler(app)
+        smc = ServiceMetadataCache(app, handler)
+        smc.start()
